requirements.txt
django==1.10
Pillow
sendgrid
django-allauth
twilio
stripe
psycopg2
